[arccore+arcane] Move and rename 'arcane/utils/Numeric.h' to 'arccore/base/TypeEqual.h'
Also remove the typedef 'TypeEqual::_IsFloatType' because this may be a
reserved keyword in C++.
